Name the end-of-game specs after the file they test
magic-box and tictactoe-alikes split into a spec per variant, next to the
module exporting the moves each exercises.
